Implement resizable T[]

+++ frameworks/bridge/arkts_frontend/koala_projects/arkoala-arkts/arkui-ohos/src/component/canvas.ts
@@ -410,7 +410,7 @@ export class ImageData implements MaterializedBase {
     }
     private getData_serialize(): NativeBuffer {
         const retval  = ArkUIGeneratedNativeModule._ImageData_getData(this.peer!.ptr)
-        return new Deserializer(retval, retval.length).readBuffer()
+        return new Deserializer(retval, retval.length as int32).readBuffer()
     }
     private getHeight_serialize(): number {
         const retval  = ArkUIGeneratedNativeModule._ImageData_getHeight(this.peer!.ptr)
@@ -1353,7 +1353,7 @@ export class CanvasRenderer extends CanvasPath implements MaterializedBase {
     }
     private getLineDash_serialize(): Array<number> {
         const retval  = ArkUIGeneratedNativeModule._CanvasRenderer_getLineDash(this.peer!.ptr)
-        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length)
+        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length as int32)
         const buffer_length : int32 = retvalDeserializer.readInt32()
         let buffer : Array<number> = new Array<number>(buffer_length)
         for (let buffer_i = 0; buffer_i < buffer_length; buffer_i++) {
@@ -1401,7 +1401,7 @@ export class CanvasRenderer extends CanvasPath implements MaterializedBase {
     }
     private measureText_serialize(text: string): TextMetrics {
         const retval  = ArkUIGeneratedNativeModule._CanvasRenderer_measureText(this.peer!.ptr, text)
-        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length)
+        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length as int32)
         const returnResult : TextMetrics = retvalDeserializer.readTextMetrics()
         return returnResult
     }
@@ -2049,7 +2049,7 @@ export class DrawingRenderingContext implements MaterializedBase {
     }
     private getSize_serialize(): Size {
         const retval  = ArkUIGeneratedNativeModule._DrawingRenderingContext_getSize(this.peer!.ptr)
-        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length)
+        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length as int32)
         const returnResult : Size = retvalDeserializer.readSize()
         return returnResult
     }
