+++ frameworks/bridge/arkts_frontend/koala_projects/arkoala-arkts/arkui-ohos/src/component/xcomponent.ts
@@ -141,7 +141,7 @@ export class XComponentController implements MaterializedBase {
     }
     private getXComponentSurfaceRect_serialize(): SurfaceRect {
         const retval  = ArkUIGeneratedNativeModule._XComponentController_getXComponentSurfaceRect(this.peer!.ptr)
-        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length)
+        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length as int32)
         const returnResult : SurfaceRect = retvalDeserializer.readSurfaceRect()
         return returnResult
     }
@@ -153,7 +153,7 @@ export class XComponentController implements MaterializedBase {
     }
     private getXComponentSurfaceRotation_serialize(): SurfaceRotationOptions {
         const retval  = ArkUIGeneratedNativeModule._XComponentController_getXComponentSurfaceRotation(this.peer!.ptr)
-        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length)
+        let retvalDeserializer : Deserializer = new Deserializer(retval, retval.length as int32)
         const returnResult : SurfaceRotationOptions = retvalDeserializer.readSurfaceRotationOptions()
         return returnResult
     }
